Move http server to own module
